Update index.js
The JavaScript code has been modified to include the textColor and textSize variables, which are initially set to white and 30, respectively. The changeTextColor() and changeTextSize() functions update these variables and call drawMeme() to redraw the meme.

The drawMeme() function has been modified to use the textColor and textSize variables to set the fill and stroke styles and font size for the text. The strokeText() and fillText() methods are called with the updated text and text position parameters.